Repackaged into .api, .impl, .main
