+++ hello-sc3/hisc3/build.sbt
@@ -1,0 +1,12 @@
+val scala3Version = "3.3.0"
+
+lazy val root = project
+  .in(file("."))
+  .settings(
+    name := "hisc3",
+    version := "0.1.0-SNAPSHOT",
+
+    scalaVersion := scala3Version,
+
+    libraryDependencies += "org.scalameta" %% "munit" % "0.7.29" % Test
+  )
