+++ hello-sc3/hisc3/project/build.properties
@@ -1,0 +1,1 @@
+sbt.version=1.9.4
