+++ hello-sc3/hisc3/src/main/scala/Main.scala
@@ -1,0 +1,5 @@
+@main def hello: Unit =
+  println("Hello world!")
+  println(msg)
+
+def msg = "I was compiled by Scala 3. :)"
