+++ hello-sc3/hisc3/src/main/scala/com/appstract/hisc3/RunMeClassic.scala
@@ -1,0 +1,8 @@
+package com.appstract.hisc3
+
+object RunMeClassic {
+	def main(args: Array[String]): Unit = {
+		println("RunMeClassic.main BEGIN")
+		println("RunMeClassic.main END")
+	}
+}
